+++ dev/com.ibm.ws.microprofile.faulttolerance.1.1.cdi/bnd.bnd
@@ -23,12 +23,11 @@ Import-Package: \
 	org.eclipse.microprofile.config.*;version="[1.0,2)",\
 	*
 
-Export-Package: \
-	com.ibm.ws.microprofile.faulttolerance.cdi
+Export-Package: com.ibm.ws.microprofile.faulttolerance11.cdi
 
 Private-Package: \
-    com.ibm.ws.microprofile.faulttolerance.cdi.resources,\
-    com.ibm.ws.microprofile.faulttolerance.cdi.config
+	com.ibm.ws.microprofile.faulttolerance.cdi.resources,\
+	com.ibm.ws.microprofile.faulttolerance11.cdi.config
 
 Include-Resource: \
     META-INF=resources/META-INF
@@ -47,8 +46,8 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 	com.ibm.websphere.javaee.cdi.1.2;version=latest,\
 	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
 	com.ibm.ws.cdi.interfaces;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.faulttolerance.1.0;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.config.1.1;version=latest,\
+	com.ibm.websphere.org.eclipse.microprofile.faulttolerance.1.1;version=latest,\
+	com.ibm.websphere.org.eclipse.microprofile.config.1.3;version=latest,\
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
         com.ibm.websphere.javaee.interceptor.1.2;version=latest,\
         com.ibm.websphere.javaee.annotation.1.2;version=latest,\
