+++ dev/com.ibm.ws.microprofile.faulttolerance.1.1.cdi/src/com/ibm/ws/microprofile/faulttolerance11/cdi/config/BulkheadConfig.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.microprofile.faulttolerance.cdi.config;
+package com.ibm.ws.microprofile.faulttolerance11.cdi.config;
 
 import java.lang.reflect.Method;
